+++ main.go
@@ -12,7 +12,6 @@ import (
 func main() {
 	database.Connect()
 	r := gin.Default()
-	// govno
 	//r.Use(cors.New(cors.Config{
 	//	AllowOrigins:     []string{"http://localhost:5500"},                 // Разрешите фронтенд
 	//	AllowMethods:     []string{"GET", "POST", "PUT", "DELETE", "PATCH"}, // Разрешенные методы
@@ -25,69 +24,71 @@ func main() {
 	// Группа для работы с хостами (Hosts)
 	hostRoutes := r.Group("/api/v1/host")
 	{
-		// Создание хоста
-		// POST /host
+		// POST /api/v1/host - Создать хост
 		hostRoutes.POST("/", controllers.CreateHost)
 
-		// Получение всех хостов
-		// GET /host
+		// GET /api/v1/host - Получить все хосты
 		hostRoutes.GET("/", controllers.GetAllHost)
 
-		// Получение хоста по IP (query-параметр ip=...)
-		// GET /host/search?ip=1.2.3.4
+		// GET /api/v1/host/search?ip=1.2.3.4 - Найти хост по IP
 		hostRoutes.GET("/search", controllers.GetHostByID)
 
-		// Обновление хоста по ID
-		// PUT /host/:id
+		// PUT /api/v1/host/:id - Обновить хост по ID
 		hostRoutes.PUT("/:id", controllers.UpdateHost)
 
-		// Удаление хоста по ID
-		// DELETE /host/:id
+		// DELETE /api/v1/host/:id - Удалить хост по ID
 		hostRoutes.DELETE("/:id", controllers.DeleteHost)
 	}
 
-	// Эндпоинт для добавления хостов в группы
-	// POST /add-hosts-to-groups
-	r.POST("/add-hosts-to-groups", controllers.AddHostToGroupHandler)
-	// Group endpoints
-
+	// Эндпоинты для работы с группами (Groups)
 	groupRoutes := r.Group("/api/v1/group")
 	{
-		// Создание группы
+		// POST /api/v1/group - Создать группу
 		groupRoutes.POST("/", controllers.CreateGroup)
 
-		// Получение всех групп
+		// GET /api/v1/group - Получить все группы
 		groupRoutes.GET("/", controllers.GetAllGroup)
 
-		// Получение группы по имени (через query-параметр "group=...")
-		// Пример: GET /group/search?group=GroupName
+		// GET /api/v1/group/search?group=GroupName - Найти группу по имени
 		groupRoutes.GET("/search", controllers.GetGroupByName)
 
-		// Обновление группы по ID (прокидываем id как /group/:id)
+		// PUT /api/v1/group/:id - Обновить группу по ID
 		groupRoutes.PUT("/:id", controllers.UpdateGroup)
 
-		// Удаление группы по ID (прокидываем id как /group/:id)
+		// DELETE /api/v1/group/:id - Удалить группу по ID
 		groupRoutes.DELETE("/:id", controllers.DeleteGroup)
 	}
 
-	// other endpoints nmap
+	// POST /api/v1/add-hosts-to-groups - Добавить хосты в группы
+	r.POST("/api/v1/add-hosts-to-groups", controllers.AddHostToGroup)
 
-	r.GET("/api/v1/pentest/:number_task", controllers.GetPentestJsonByNumberTask)
-	r.POST("/api/v1/upload-script", controllers.UploadScript) // Загрузить скрипт nmap на сервер
-	r.POST("/api/v1/nmap", controllers.ProcessNmapRequest)
+	// Эндпоинты для работы с задачами (Tasks)
+
+	// POST /api/v1/task - Создать задачу
+	r.POST("/api/v1/task", controllers.CreateTask)
+
+	// GET /api/v1/task-status/:number_task - Проверить статус задачи
 	r.GET("/api/v1/task-status/:number_task", controllers.GetTaskStatus)
+
+	// DELETE /api/v1/task/:number_task - Удалить задачу
 	r.DELETE("/api/v1/task/:number_task", controllers.DeleteTask)
+
+	// GET /api/v1/task-all - Получить все задачи
 	r.GET("/api/v1/task-all", controllers.GetTaskAll)
 
-	// Получение результатов Nmap
-	r.GET("/api/v1/last-nmap", controllers.GetLastNmap)
-	r.GET("/api/v1/all-nmap", controllers.GetAllNmap)
-	//r.GET("/api/v1/name-nmap/:filename", controllers.GetReportByName)
+	// Остальные эндпоинты
 
-	// Загружаем 2 файла БДУ ФСТЭКА, сохраняем и обрабатываем
-	// Не трогать, тут как часики работает
-	r.POST("/api/v1/update-database-bdu", controllers.UploadDatabaseBDU)
+	// POST /api/v1/upload-script - Загрузить скрипт Nmap
+	r.POST("/api/v1/upload-script", controllers.UploadScript)
+
+	// GET /api/v1/pentest/:number_task - Получить отчет пентеста
+	r.GET("/api/v1/pentest/:number_task", controllers.GetPentestJsonByNumberTask)
+
+	// WebSocket подключение
+	// GET /api/v1/ws - Подключение WebSocket
 	r.GET("/api/v1/ws", controllers.HandleWebSocket)
+
+	// Запуск сервера на порту 3001
 	r.Run(":3001")
 
 }
